remove input files from git
Remove the full input files from Git and GitHub. Users are advised not to share the input, which I did not know of. They are now removed from future versions of this repository (altough they can be found in the history). I hope I am forgiven for sharing them without knowing that it is advised not to, since I did not know.